+++ src/components/MultiSelect/CHANGELOG.md
@@ -1,1 +1,2 @@
 - 2025-11-06: Initial version.
+- 2025-12-16: Use `IbsBase` as the base component, etc.